move evalautor_utils to solve circular dependency issue
