+++ ragelo/evaluators/__init__.py
@@ -1,5 +1,4 @@
 from ragelo.evaluators.answer_evaluators import PairwiseAnswerEvaluator
-from ragelo.evaluators.evaluator_utils import resolve_evaluator_result_type
 from ragelo.evaluators.retrieval_evaluators import (
     CustomPromptEvaluator,
     DomainExpertEvaluator,
@@ -8,6 +7,7 @@ from ragelo.evaluators.retrieval_evaluators import (
     ReasonerEvaluator,
     RetrievalEvaluatorFactory,
 )
+from ragelo.types.evaluator_utils import resolve_evaluator_result_type
 
 __all__ = [
     "PairwiseAnswerEvaluator",
